Fix problem with original Post entity
Problem occurred because I was using the .HasForeignKey() method when setting up the Blog properties. I had fixed this by adding a Blog property to the Post entity, but I wanted to remove it and try the original way. The use of the .HasForeignKey() method was unnecessarily adding an additional key reference and causing problems.